fix: resolve all compiler warnings
- typed_mailbox.hpp: initialize missing TimsHeader fields (dest, src, flags)
  in both send() overloads
- tims_helpers.hpp: initialize missing TimsHeader fields (dest, src, flags)
- module_output.hpp: initialize missing TimsHeader fields (msg_size, dest, src, flags)
- continuous_input.hpp: initialize missing TimsHeader fields (msg_size, dest, flags)
  in UnsubscribeRequest
- synced_input.hpp: initialize missing TimsHeader fields (msg_size, flags)
  in GetDataRequest and GetNextDataRequest
- module2.hpp: remove unused 'result' variable from receive_any_for()
- test_module_output_getdata.cpp: fix narrowing int -> uint32_t conversion
- test_multi_input_compilation.cpp: suppress unused parameter warnings with (void)
- test_process_signature.cpp: suppress unused local typedef warnings with sizeof()
- test_type_name.cpp: use anonymous parameters in main() signature
- test_auto_id_validation.cpp: suppress unused typedef warning with sizeof()

+++ test/test_module_output_getdata.cpp
@@ -226,8 +226,8 @@ void test_get_data_wrap_around() {
     
     // Now publish more data to have t1 as oldest in buffer and full buffer (10 messages)
     for (int i = 0; i < 7; ++i) {
-        output.publish(SensorData{.value = 40.0f + i,
-            .sensor_id = 4 + i}, t3 + (i + 1) * 1000000000ULL);
+        output.publish(SensorData{.value = 40.0f + static_cast<float>(i),
+            .sensor_id = static_cast<uint32_t>(4 + i)}, t3 + (i + 1) * 1000000000ULL);
     }
     reply = output.handle_get_data_request(req);
